[#1] Refactor: fetch 로직 api/fetchTx.ts 로 리펙토링
